Сделал отдельным компонентом блок карточек
