+++ src/pages/main/components/CoinsContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CoinsContainer";
