+++ src/pages/main/components/ListCoins/ListCoins.tsx
@@ -1,76 +1,47 @@
 import React from "react";
 
 import Card from "@components/Card";
-import Pagination from "@components/Pagination";
-import WithLoader from "@components/WithLoader";
-import { useLocalStore } from "@hooks/useLocaleStore";
-import CoinsStore from "@store/CoinsStore";
-import rootStore from "@store/RootStore/instance";
-import convertNumberToArray from "@utils/convertNumberToArray";
-import { Meta } from "@utils/meta";
-import { observer } from "mobx-react-lite";
-import { useNavigate, useSearchParams } from "react-router-dom";
+import { CoinsModel } from "@store/models/coins";
+import { useNavigate } from "react-router-dom";
 
 import styles from "./ListCoins.module.scss";
 import Price from "../Price";
 
-const ListCoins = () => {
-  const coinsStore = useLocalStore(() => new CoinsStore());
+/** Пропсы, которые принимает компонент ListCoins */
+export type ListCoinsProps = {
+  /** Криптомонеты */
+  items: CoinsModel[];
+};
 
-  const [searchParams, setSearchParams] = useSearchParams();
-
+const ListCoins: React.FC<ListCoinsProps> = ({ items }: ListCoinsProps) => {
   const navigate = useNavigate();
-
-  const setPage = (number: number) => {
-    setSearchParams({ ...rootStore.query.params, page: String(number) });
-  };
-
-  React.useEffect(() => {
-    if (!rootStore.query.params.page) {
-      setSearchParams({ page: "1", ...rootStore.query.params });
-    } else {
-      coinsStore.getCoins({
-        ...rootStore.query.params,
-        page: rootStore.query.params.page,
-      });
-    }
-  }, [coinsStore, rootStore.query.params]);
 
   const goToPage = (id: string): void => {
     navigate(`/coins/${id}`);
   };
 
   return (
-    <WithLoader loading={coinsStore.loading === Meta.loading}>
-      <div className={styles.listCoins}>
-        {coinsStore.items.map((item) => {
-          return (
-            <React.Fragment key={item.id}>
-              <Card
-                image={item.image}
-                title={item.name}
-                subtitle={item.symbol}
-                content={
-                  <Price
-                    price={item.currentPrice}
-                    percentage={item.priceChangePercentage}
-                  />
-                }
-                onClick={() => goToPage(item.id)}
-              />
-            </React.Fragment>
-          );
-        })}
-      </div>
-      <Pagination
-        currentPage={Number(rootStore.query.params.page)}
-        pages={convertNumberToArray(
-          coinsStore.totalCount / coinsStore.items.length
-        )}
-        loadingPage={setPage}
-      />
-    </WithLoader>
+    <div className={styles.listCoins}>
+      {items.map((item) => {
+        return (
+          <React.Fragment key={item.id}>
+            <Card
+              image={item.image}
+              title={item.name}
+              subtitle={item.symbol}
+              content={
+                <Price
+                  price={item.currentPrice}
+                  percentage={item.priceChangePercentage}
+                />
+              }
+              onClick={() => goToPage(item.id)}
+            />
+          </React.Fragment>
+        );
+      })}
+    </div>
   );
 };
 
-export default observer(ListCoins);
+export default ListCoins;
